requirements.txt
numpy
pandas
pyarrow 
python-dotenv
scikit-learn
requests
plotly
Flask-WTF
openpyxl
waitress
pytz
flask-cors
gevent
joblib
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug